+++ packages/azure/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @ai-sdk/azure
 
+## 4.0.18
+
+### Patch Changes
+
+- Updated dependencies [02ffdcb]
+- Updated dependencies [76cb673]
+  - @ai-sdk/provider-utils@5.0.12
+  - @ai-sdk/deepseek@3.0.13
+  - @ai-sdk/openai@4.0.17
+
 ## 4.0.17
 
 ### Patch Changes