+++ packages/llamaindex/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ai-sdk/llamaindex
 
+## 3.0.34
+
+### Patch Changes
+
+- ai@7.0.34
+
+## 3.0.33
+
+### Patch Changes
+
+- Updated dependencies [76cb673]
+- Updated dependencies [e808fa5]
+- Updated dependencies [33647d7]
+  - ai@7.0.33
+
 ## 3.0.32
 
 ### Patch Changes